Add ApproxPipelineSpec add license header Java class path maybe too long when invoking sbt.ForkMain? update jvmopts
