+++ .jvmopts
@@ -6,6 +6,7 @@
 -XX:ReservedCodeCacheSize=512m
 -XX:+TieredCompilation
 -XX:+UseParallelGC
+-XX:+CMSClassUnloadingEnabled
 
 # https://github.com/spotify/scio/issues/2096
 # https://github.com/sbt/sbt/issues/4794#issuecomment-500602546
